+++ bin/gmaproto.tcl
@@ -1,12 +1,12 @@
 ########################################################################################
-#  _______  _______  _______                ___        __     _____      _______       #
-# (  ____ \(       )(  ___  ) Game         /   )      /  \   / ___ \    / ___   )      #
-# | (    \/| () () || (   ) | Master's    / /) |      \/) ) ( (   ) )   \/   )  |      #
-# | |      | || || || (___) | Assistant  / (_) (_       | | ( (___) |       /   )      #
-# | | ____ | |(_)| ||  ___  |           (____   _)      | |  \____  |     _/   /       #
-# | | \_  )| |   | || (   ) |                ) (        | |       ) |    /   _/        #
-# | (___) || )   ( || )   ( | Mapper         | |   _  __) (_/\____) ) _ (   (__/\      #
-# (_______)|/     \||/     \| Client         (_)  (_) \____/\______/ (_)\_______/      #
+#  _______  _______  _______                ___       _______  _______                 #
+# (  ____ \(       )(  ___  ) Game         /   )     / ___   )(  __   )                #
+# | (    \/| () () || (   ) | Master's    / /) |     \/   )  || (  )  |                #
+# | |      | || || || (___) | Assistant  / (_) (_        /   )| | /   |                #
+# | | ____ | |(_)| ||  ___  |           (____   _)     _/   / | (/ /) |                #
+# | | \_  )| |   | || (   ) |                ) (      /   _/  |   / | |                #
+# | (___) || )   ( || )   ( | Mapper         | |   _ (   (__/\|  (__) |                #
+# (_______)|/     \||/     \| Client         (_)  (_)\_______/(_______)                #
 #                                                                                      #
 ########################################################################################
 #
@@ -57,9 +57,9 @@ package require base64 2.4.2
 package require uuid 1.0.1
 
 namespace eval ::gmaproto {
-	variable protocol 409
+	variable protocol 410
 	variable min_protocol 333
-	variable max_protocol 409
+	variable max_protocol 410
 	variable max_max_protocol 499
 	variable debug_f {}
 	variable legacy false
@@ -82,6 +82,7 @@ namespace eval ::gmaproto {
 	variable current_stream {}
 	variable stream_dict {}
 	variable progress_stack {}
+	variable ClientSettings {}
 
 	variable _message_map
 	array set _message_map {
@@ -166,13 +167,14 @@ namespace eval ::gmaproto {
 		PROGRESS {OperationID s Title s Value i MaxValue i IsDone ?}
 		PS      {ID s Name s Health {o {MaxHP i LethalDamage i NonLethalDamage i Con i IsFlatFooted ? IsStable ? Condition s HPBlur i}} Gx f Gy f Skin i SkinSize l PolyGM ? Elev i Color s Note s Size s DispSize s StatusList l AoE {o {Radius f Color s}} MoveMode i Reach i Killed ? Dim ? CreatureType i Hidden ? CustomReach {o {Enabled ? Natural i Extended i}}}
 		READY   {}
+		REDIRECT {Host s Port i Reason s}
 		ROLL    {Sender s Recipients l MessageID i ToAll ? ToGM ? Title s Result {o {InvalidRequest ? ResultSuppressed ? Result i Details {a {Type s Value s}}}} RequestID s MoreResults ?}
 		SYNC    {}
 		SYNC-CHAT {Target i}
 		TB      {Enabled ?}
 		TO      {Sender s Recipients l MessageID i ToAll ? ToGM ? Text s}
 		UPDATES {Packages {a {Name s Instances {a {OS s Arch s Version s Token s}}}}}
-		WORLD   {Calendar s}
+		WORLD   {Calendar s ClientSettings {o {MkdirPath s ImageBaseURL s ModuleCode s SCPDestination s ServerHostname s}}}
 		/CONN   {}
 		Animation {Frames i FrameSpeed i Loops i}
 		Health  {MaxHP i LethalDamage i NonLethalDamage i Con i IsFlatFooted ? IsStable ? Condition s HPBlur i}
@@ -301,7 +303,7 @@ proc ::gmaproto::redial {} {
 	after 10 ::gmaproto::_receive $::gmaproto::sock
 
 	if [catch {::gmaproto::_login} err] {
-		say "Attempt to sign on to server failed: $err"
+		::say "Attempt to sign on to server failed: $err"
 	}
 }
 
@@ -1199,6 +1201,12 @@ proc ::gmaproto::int_bool {b} {
 		return 0
 	}
 }
+
+# unfortunately, the tcllib JSON interpreter is broken and cannot distinguish
+# between a JSON null value and a string "null". So we are forced to treat a
+# null value (including a string "null") as a null. So you can't have any string
+# with the value of "null". 
+# Sorry.
 proc ::gmaproto::_construct {input types} {
 	foreach {field t} $types {
 		switch -exact -- [lindex $t 0] {
@@ -1312,10 +1320,13 @@ proc ::gmaproto::_construct {input types} {
 				}
 			}
 			d {
-				if {[dict exists $input $field]} {
-					if {[dict get $input $field] eq "null"} {
-						dict set input $field {}
-					}
+				if {[dict exists $input $field] && [dict get $input $field] ne "null"} {
+					dict set input $field [dict map {k v} [dict get $input $field] {
+						if {$v eq "null"} {
+							set v {}
+						}
+						set v
+					}]
 				} else {
 					dict set input $field {}
 				}
@@ -1920,8 +1931,29 @@ proc ::gmaproto::_login {} {
 			}
 			AC	{ ::gmaproto::_dispatch_to_app AC $params }
 			DSM	{ ::gmaproto::_dispatch_to_app DSM $params }
+			REDIRECT {
+				::gmautil::dassign $params Host newhost Port newport Reason reason
+				::INFO "Server ${::gmaproto::host}:${::gmaproto::port} asked us to connect instead to the server at ${newhost}:${newport}"
+				if {$reason ne {}} {
+					::INFO "Redirecting because $reason"
+				}
+				set ::gmaproto::sock {}
+				set ::gmaproto::pending_login true
+				set ::gmaproto::host $newhost
+				set ::gmaproto::port $newport
+				catch {close $::gmaproto::sock}
+				set ::gmaproto::send_buffer {}
+				set ::gmaproto::recv_buffer {}
+				set ::gmaproto::read_buffer {}
+				set ::gmaproto::poll_buffer {}
+				::gmaproto::redial
+				return
+			}
 			MARCO	{ ::gmaproto::DEBUG "Ignored MARCO during login" }
-			WORLD	{ set calendar [dict get $params Calendar] }
+			WORLD	{ 
+				set calendar [dict get $params Calendar] ;# TODO currently not used; the mapper only recognizes the Golarion calendar
+				::gmaproto::_dispatch_to_app WORLD $params
+			}
 			DENIED {
 				::report_progress "Server denied access"
 				::say "Server DENIED access: [dict get $params Reason]"
@@ -2232,7 +2264,7 @@ proc ::gmaproto::normalize_dict {cmd d} {
 	return [::gmaproto::new_dict_from_json $cmd [::gmaproto::json_from_dict $cmd $d]]
 }
 
-# @[00]@| GMA-Mapper 4.19.2
+# @[00]@| GMA-Mapper 4.20
 # @[01]@|
 # @[10]@| Copyright © 1992–2023 by Steven L. Willoughby (AKA MadScienceZone)
 # @[11]@| steve@madscience.zone (previously AKA Software Alchemy),